test(tui): split settings screen tests
Move the settings screen tests into focused submodules so the refactor remains within the Rust file-size gate.

Refs #398
